fix routing to login page when clicking on private room.
